requirements.txt
pyrogram
pyrogram[asyncio]
python-dotenv
setuptools
llama-cpp-python
emoji